+++ example/android/app/src/main/kotlin/dev/hebei/clover_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.hebei.clover_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()