+++ example/lib/view_model.dart
@@ -1,1 +1,0 @@
-export 'view_model/home_view_model.dart';