+++ example/lib/views/home_view.dart
@@ -1,0 +1,37 @@
+import 'package:clover/clover.dart';
+import 'package:clover_example/view_models.dart';
+import 'package:flutter/material.dart';
+
+/// The home view.
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = ViewModel.of<HomeViewModel>(context);
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(viewModel.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'You have pushed the button this many times:',
+            ),
+            Text(
+              '${viewModel.count}',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () => viewModel.increase(),
+        tooltip: 'Increse',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}